Mejorando las tabs dando mejor intuicion y diseño
