+++ app/(tabs)/Home/home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Landing from './components/Landing';
+import Landing from '../../components/Landing';
 
 export default function Home() {
     return (
