+++ app/(tabs)/Home/sectionh.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Sections from './components/Sections';
+import Sections from '../../components/Sections';
 
 export default function Sectionh() {
     return (
