+++ app/_layout.js
@@ -1,13 +1,15 @@
-import { Slot } from "expo-router";
+import { Stack } from "expo-router";
 import { StatusBar } from "expo-status-bar";
 import React from "react";
 import { View } from "react-native";
 
 export default function Layout() {
     return (
-        <View style={{backgroundColor: '#8A7AE7'}}>
-            <StatusBar style="light" />
-            <Slot />
+        <View style={{ flex: 1 }}>
+            <StatusBar style="dark" />
+            <Stack>
+                <Stack.Screen name="(tabs)" options={{headerShown: false}} />
+            </Stack>
         </View>
     );
 }
