+++ components/Works.jsx
@@ -1,8 +1,10 @@
 import { Link } from 'expo-router';
 import React from 'react';
+import { useRouter } from "expo-router"; 
 import { View, Pressable, Text, StyleSheet, ScrollView, Dimensions } from 'react-native';
 const { width, height } = Dimensions.get('window');
 export default function Works ()  {
+    const router = useRouter();
     return (
         <View style={styles.all}>
             <ScrollView>
@@ -14,7 +16,7 @@ export default function Works ()  {
                     </View>
                     <View style={styles.container}>
                         <Link asChild href="#">
-                            <Pressable onPress={() => alert('Lengua')}>
+                            <Pressable onPress={() => router.push(`/id`)}>
                                 <Text style={styles.text}>Lengua</Text>
                             </Pressable>
                         </Link>
@@ -99,6 +101,7 @@ const styles = StyleSheet.create({
         flexDirection: 'column',
         justifyContent: 'space-between',
         padding: 12,
+        backgroundColor: '#8A7AE7'
     },
     container: {
         flex: 1,
@@ -106,7 +109,7 @@ const styles = StyleSheet.create({
         width: width - 50,
         height: height / 10,
         alignItems: 'center',
-        backgroundColor: 'rgb(255, 255, 255)',
+        backgroundColor: 'rgba(255, 255, 255, 0.61)',
         margin: 10,
         borderRadius: 12,
     },
