+++ package.json
@@ -21,7 +21,7 @@
     "expo-status-bar": "~2.0.1",
     "react": "^18.3.1",
     "react-native": "^0.76.6",
-    "react-native-safe-area-context": "^4.12.0",
+    "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0"
   },
   "devDependencies": {
